refactor birdnet based models into separate files
